requirements.txt
fastapi==0.118.0
uvicorn==0.23.2
pydantic==2.8.0
python-dotenv==1.0.0
langchain==1.2.0
langchain-core==0.3.0
langchain-openai==1.1.3
langchain-community==0.4.1
tiktoken==0.12.0
psycopg2-binary==2.9.8
sqlalchemy==2.0.45